% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{Basis}
\alias{Basis}
\title{Basis}
\description{
This class creates a basis of length d, which generates a d-dimensional linear subspace of \eqn{L^2[0,1]}. Specifically, 
a trigonometric basis and a histogramm basis are provided as options.
}
\details{
Note that the parent class \code{Basis} is merely there for construction purposes. An object generated by 
\eqn{Basis$new(d)} is not of much use to the user. Therefore one should always generat an object of a subclass of \code{Basis},
i.e. an object of class \code{Trig_Basis} or \code{Hist_Basis}
}
\section{Usage}{

\preformatted{trig_bas <- Trig_Basis$new(d)
hist_bas <- Hist_Basis$new(d)


trig_bas$get_function(j)
trig_bas$dimension
hist_bas$get_function(j)
hist_bas$dimension
}
}

\section{Arguments}{

\describe{
  \item{d}{a natural number which specifies the dimension of the subsapce generated by the object of class basis.}
  \item{j}{a natural number smaller or equal to \code{d} which specifies which basis function is returned.}  
}
}

\section{Methods}{

\code{$new()} generates a new \code{Basis} object dependent on the input \code{d}.

\code{$get_function()} returns the jth basis function \code{j}.

\code{$dimension} returns the dimension of the vector space generated by the object of the class.
}

\examples{
trig_bas <- Trig_Basis$new(10)
trig_bas$get_function(4)
trig_bas

hist_bas <- Hist_Basis$new(10)
hist_bas$get_function(5)
hist_bas
}
