% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datsim.R
\name{datsim}
\alias{datsim}
\title{Simulation of data}
\usage{
datsim(dens, m, n)
}
\arguments{
\item{dens}{a probability density on \eqn{[0,1]}}

\item{m}{a regression function}

\item{n}{a numeric value, specifying the number of pairs of values to be generated}
}
\value{
an \code{n} x 2-data.frame
}
\description{
Creates a data frame consisting of pairs (x, y) where the first component is distributed
according to a given probability density and y satisfies y = m(x) + e where e is a normally distributed error term.
}
\examples{
#define a density on [0,1]
dens1 <- function(x) 1 + 0.5 * sin(2 * pi * x)
#using the log-function as regression function
datasim1 <- datsim(dens1, log, 50)
datasim1
}
