% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_dens.R
\name{est_dens}
\alias{est_dens}
\title{One dimensional projection estimator}
\usage{
est_dens(basis, data, D)
}
\arguments{
\item{basis}{an object of class \code{\link{Basis}}}

\item{data}{a numeric vector containing a sample of i.i.d. real random variables with common density}

\item{D}{a natural number smaller or equal to the dimension of the vector space generated by \code{basis}}
}
\value{
The projection estimator of the density of \code{data} given \code{basis}
}
\description{
Given a sample of i.i.d. real random variables with common density, and a basis of class \code{\link{Basis}}
this function calculates the projection estimator of this density.
}
\examples{
# example with the trigonometric basis
d <- 100
trig_bas <- Trig_Basis$new(d)
data <- runif(100)
estimated_density <- est_dens(trig_bas, data, 10)
estimated_density
}
