% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_dens2.R
\name{est_dens2}
\alias{est_dens2}
\title{Two dimensional projection estimator}
\usage{
est_dens2(basis, data, dim)
}
\arguments{
\item{basis}{an object of class \code{\link{Basis}}}

\item{data}{a data frame with two columns of samples of i.i.d. real random variables}

\item{dim}{a vector of length two. If only one value is passed, this value is reproduced to attain a vector of length two}
}
\value{
The two dimensional projection estimator of the common density of the data stored in \code{data}, given \code{basis}
}
\description{
Given a sample of i.i.d real random bivariate vectors with a common density f and a basis of
class \code{\link{Basis}}. This function calculates the two dimensional projection estimator of the destiny f.
}
\examples{
# example with the trigonometric basis
d <- 10
d1 <- 8
d2 <- 5
trig_bas <- Trig_Basis$new(d)
x <- runif(100)
y <- runif(100)
estimated_density <- est_dens2(trig_bas, data = data.frame(x, y), dim = c(d1, d2))
}
