% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_dens_squared.R
\name{est_dens_squared}
\alias{est_dens_squared}
\title{Square projection estimator}
\usage{
est_dens_squared(basis, data, D)
}
\arguments{
\item{basis}{an object of class \code{\link{Basis}}}

\item{data}{a sample of i.i.d. real random variables with common density}

\item{D}{a natural number smaller or equal to the Dimension generated by \code{basis}}
}
\value{
the squared projection estimator of the underlying density of \code{data}
}
\description{
this function squares the density estimated by the projection estimator
}
\keyword{internal}
