% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_proj2_coef.R
\name{est_proj2_coef}
\alias{est_proj2_coef}
\title{two dimensional projection estimator coefficients}
\usage{
est_proj2_coef(basis, data1, data2, j, k)
}
\arguments{
\item{basis}{an object of class \code{\link{Basis}}}

\item{data1}{a sample of i.i.d. real random variables}

\item{data2}{a sample of i.i.d. real random variables with the same length as \code{data1}}

\item{j}{a natural number between 1 and the dimension of \code{basis}}

\item{k}{a natural number between 1 and the dimension of \code{basis}}
}
\value{
the (\code{j}, \code{k})-th estimated coefficient of the estimated density based on \code{basis}, \code{data1} and \code{data2}
}
\description{
Given a sample of i.i.d. real random bivariate vectors with a common density f and a basis of class \code{\link{Basis}}. 
This function calculates the coefficients of the two dimensional projection estimator of the density f.
}
\keyword{internal}
