% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_proj_coef.R
\name{est_proj_coef}
\alias{est_proj_coef}
\title{One dimensional projection estimator coefficients}
\usage{
est_proj_coef(basis, data, j)
}
\arguments{
\item{basis}{an object of class \code{\link{Basis}}}

\item{data}{a sample of i.i.d. real random variables with common density}

\item{j}{a natural number between 1 and the dimension of \code{basis}}
}
\value{
the jth estimated coefficient of the estimated density based on the \code{basis} and \code{data}
}
\description{
Given a sample of i.i.d. real rand. variables with a common density f and a basis of class \code{\link{Basis}}. 
This function calculates the coefficients of the projection estimator of the density f.
}
\keyword{internal}
