% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\docType{data}
\name{Hist_Basis}
\alias{Hist_Basis}
\title{Histogram Basis}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Hist_Basis
}
\description{
This class creates a trigonometric basis of length d, which generates a d-dimensional linear subspace of \eqn{L^2[0,1]}
}
\details{
For a detailed documentation please see the documentation of class \code{\link{Basis}}, which is the parent class.
}
\keyword{datasets}
