% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midpoint_composite.R
\name{midpoint_composite}
\alias{midpoint_composite}
\title{Midpoint Composite rule}
\usage{
midpoint_composite(f, a, b, n)
}
\arguments{
\item{f}{a function}

\item{a}{the lower bound}

\item{b}{the upper bound}

\item{n}{the amount of intervals we divide the interval [\code{a}, \code{b}] into}
}
\value{
an estimation of the integral of f over [\code{a}, \code{b}]
}
\description{
This function implements the midpoint composite rule for numerical integration
}
\keyword{internal}
