% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba-data.R
\docType{data}
\name{nba}
\alias{nba}
\title{Nba Data}
\format{An object of class \code{"data.frame"}.}
\source{
\href{http://www.basketball-reference.com/}{QTL Archive} \href{http://stats.nba.com/}{QTL Archive}
}
\usage{
data(nba)
}
\description{
Data about performance statistics and salaries of 290 Nba players during the 2016-17 season
}
\keyword{datasets}
