% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfect_D.R
\name{perfect_D}
\alias{perfect_D}
\title{Adaptive estimator for the dimension}
\usage{
perfect_D(basis, data, kappa = 2)
}
\arguments{
\item{basis}{an object of class \code{\link{Basis}}}

\item{data}{a numeric vector containing a sample of i.i.d. real random variables with common density}

\item{kappa}{a numeric constant which does not depend on the data. A value greater than 1 a suitable, the default is 2}
}
\value{
a numeric value, the dimension of a suitable subspace
}
\description{
Given an orthonormal basis and a data sample, this function calculates a relevant dimension D, which represents the
dimension we want to calculate our projection estimator in (see \code{\link{est_dens}} for more details).
}
\examples{
\dontrun{
trig_bas <- Trig_Basis$new(1000)
data <- rnorm(100, 0.5, 0.088)
perfect_D(trig_bas, data)
}
}
