% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basis.R
\name{plot_basis}
\alias{plot_basis}
\title{Plotting basis functions}
\usage{
plot_basis(basis, ind, resol = 300)
}
\arguments{
\item{basis}{an object of class \code{\link{Basis}}}

\item{ind}{a numeric atomic vector, containing information about which functions of the \code{basis} should be plotted}

\item{resol}{a numeric value, specifying the quality of the plot(length.out of processed values)}
}
\value{
no return
}
\description{
Given an object of class \code{\link{Basis}}, this function plots the functions contained in this object.
}
\examples{

hist <- Hist_Basis$new(10)
plot_basis(hist, c(1, 3, 7))

trig <- Trig_Basis$new(100)
#plot the first three basis functions
plot_basis(trig, seq(3))

}
