% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dens2.R
\name{plot_dens2}
\alias{plot_dens2}
\title{Plot bivariate estimated density function}
\usage{
plot_dens2(basis, data, dim, resol = 300)
}
\arguments{
\item{basis}{an object of class \code{\link{Basis}}}

\item{data}{a data frame with two columns, each containing a sample of i.i.d. real random variables with common density.}

\item{dim}{a vector of length two. If only one value is passed, this value is reproduced to attain a vector of length two.}

\item{resol}{a numeric value, specifying the quality of the plot(length.out of processed values). The default is 300.}
}
\value{
no return
}
\description{
This function plots the estimated bivariate density function for a given set of data using a given orthonormal basis
of some dimension.
}
\examples{
\dontrun{
trig_bas <- Trig_Basis$new(1000)
dens <- function(x) dnorm(x, 0.5, 0.088)
data <- rdens(dens, 150)
p9 <- plot_dens2(trig_bas, data = data.frame(data, data), dim = c(8, 8))
}
}
