% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg.R
\name{plot_reg}
\alias{plot_reg}
\title{Plot regression estimator}
\usage{
plot_reg(basis1, basis2, data, f = NULL, compare = TRUE)
}
\arguments{
\item{basis1}{an object of class \code{\link{Basis}}}

\item{basis2}{an object of class \code{\link{Basis}}}

\item{data}{a data frame consisting of two columns. The two columns represent an observed sample of i.i.d. real random variables 
\eqn{(X_{i}, Y_{i}), 1 \le i \le n}.}

\item{f}{a function. This function (if passed) may be used to compare the estimate to a known function. The default ist NULL.}

\item{compare}{a logical value. If TRUE (the default), the regression estimate is compared to \code{f} if \code{f} is passed.
If FALSE only the estimate is plotted}
}
\value{
no return
}
\description{
This function plots the regression estimator given by \code{\link{reg_est}}.
}
\details{
Note, that the graph only contains a scatter plot consisting of the data points passed by \code{data}.
}
\examples{
\dontrun{
hist_bas <- Hist_Basis$new(15)
dens1 <- function(x) dnorm(x, 0.5, 0.088)
datasim <- datsim(dens1, log, 15)
plot_reg(hist_bas, hist_bas, datasim, f = log)
}
}
