% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdens.R
\name{rdens}
\alias{rdens}
\title{Random density deviates}
\usage{
rdens(dens, n)
}
\arguments{
\item{dens}{a probability density on the intervall \eqn{[0,1]}}

\item{n}{a natural number, which specifies the number of deviates to be generated}
}
\value{
a numeric vector of length \code{n} containing random deviates of \code{dens}
}
\description{
This function generates random deviates of a given probability density analogously to \code{runif} and \code{rnorm} 
in case of the uniform and normal distribution respectively.
}
\examples{
#defining a density on [0,1]
dens1 <- function(x) 1 + 0.5 * sin(2 * pi * x)
#generated 50 deviates
rdens(dens1, 50)
}
