% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_est.R
\name{reg_est}
\alias{reg_est}
\title{Adaptive estimator for a regression problem}
\usage{
reg_est(basis1, basis2, data)
}
\arguments{
\item{basis1}{an object of class \code{\link{Basis}}}

\item{basis2}{an object of class \code{\link{Basis}}}

\item{data}{a data frame consisting of two columns. The two columns represent an observed sample of i.i.d. real random variables 
\eqn{(X_{i}, Y_{i}), 1 \le i \le n}.}
}
\value{
An estimate for \eqn{m}
}
\description{
We observe a sample of i.i.d. real random variables \eqn{(X_{i}, Y_{i}), 1 \le i \le n} and consider the model
\eqn{Y_i = m(X_i) + \varepsilon_i}. The \eqn{\varepsilon_{i}} are i.i.d., centred, with common variance, the \eqn{X_i}
are i.i.d. with common density \eqn{f}. Moreover, the \eqn{(X_i)_{1 \le i \le n}} are independent. This function deduces an estimator of the
function m.
}
\details{
To compute the estimate of m, the quotient of the adaptive estimator of \eqn{l := m*f} and the adaptive estimator of \eqn{f}
is used (see \code{\link{est_dens}} and \code{\link{perfect_D}}). These two estimators can be calculated using two different bases.
}
\examples{
\dontrun{
trig_bas <- Trig_Basis$new(100)
dens1 <- function(x) dunif(x)
data <- datsim(den1, log, 150)
reg_m <- reg_est(trig_bas, trig_bas, data)
reg_m
}
}
